Migrate #testRemoveConnectorConfigSlowProducer to Mockito
